requirements.txt
pymupdf>=1.22.0
pdfminer.six>=20221105
pytesseract>=0.3.7
pillow>=9.0.0
watchdog>=2.0.0
